Support sending events without an `event:` field for the event name
The event field is optional in the specification, and adds overhead,
which some applications do not want. Allow omitting it.

To simplify use, accept the name as `impl Into<Option<&str>>`, which
allows existing code that passes `"name"` to continue working. Since
this can potentially cause an inference failure in code that previously
compiled, bump the major version, but *most* code should continue
working with just a recompile.