add support for multiline data messages.
Currently, the `data` passed to `Sender::send()` is prepended with
`data:` and written as-is. If the data string contains newlines, that
causes any data beyond the newline(s) to end up on a line that does not
begin with `data:`, which the spec dictates should be ignored. This
change splits the message on newlines and writes each line with its own
`data:` prefix.